authentication/requirements.txt
Flask
mysql-connector-python
PyJWT
cryptography
Flask-Limiter